test(cisco-ai-defense): split chat and mcp tests
