+++ tests/test_litellm/proxy/guardrails/guardrail_hooks/test_cisco_ai_defense_chat.py
@@ -1,253 +1,44 @@
-import json
-import os
-import sys
-from contextlib import contextmanager
-from datetime import datetime
-from types import SimpleNamespace
-from typing import Any, Dict
-from unittest.mock import AsyncMock, patch
-
-import pytest
-from fastapi import HTTPException
-from httpx import Request, Response
-
-from litellm.types.utils import (
+from tests.test_litellm.proxy.guardrails.guardrail_hooks._cisco_ai_defense_test_utils import (
+    Any,
+    AsyncMock,
+    CHAT_URL,
     Choices,
+    CiscoAIDefenseGuardrail,
+    CiscoAIDefenseGuardrailMissingSecrets,
     Delta,
+    DualCache,
+    HTTPException,
+    MCP_URL,
     Message,
     ModelResponse,
     ModelResponseStream,
+    Response,
+    SimpleNamespace,
     StreamingChoices,
+    UserAPIKeyAuth,
+    _aiter,
+    _chat_request_function_call_args,
+    _chat_request_tool_call_args,
+    _find_callback,
+    _make_guardrail,
+    _make_model_response_with_content,
+    _make_streaming_chunks,
+    _mcp_request,
+    _mcp_response,
+    _mock_inspect_response,
+    _patch_inspection_post,
+    _redact_response,
+    _responses_api_response,
+    _safe_response,
+    _streaming_setup,
+    _violation_response,
+    datetime,
+    init_guardrails_v2,
+    litellm,
+    os,
+    patch,
+    pytest,
 )
-
-
-def _make_model_response_with_content(content: str) -> ModelResponse:
-    """Build a real ModelResponse so ``isinstance(choice, Choices)`` works.
-
-    MagicMock(spec=Choices) doesn't expose Pydantic v2 field attributes, so
-    a real Choices/Message pair is the simplest way to mirror what the
-    proxy hands the post-call hook in production.
-    """
-    return ModelResponse(
-        choices=[
-            Choices(
-                index=0,
-                finish_reason="stop",
-                message=Message(role="assistant", content=content),
-            )
-        ]
-    )
-
-
-sys.path.insert(0, os.path.abspath("../.."))
-
-import litellm
-from litellm import DualCache
-from litellm.proxy._types import UserAPIKeyAuth
-from litellm.proxy.guardrails.guardrail_hooks.cisco_ai_defense import (
-    CiscoAIDefenseGuardrail,
-    CiscoAIDefenseGuardrailMissingSecrets,
-)
-from litellm.proxy.guardrails.init_guardrails import init_guardrails_v2
-
-CISCO_BASE = "https://us.api.inspect.aidefense.security.cisco.com"
-CHAT_URL = f"{CISCO_BASE}/api/v1/inspect/chat"
-MCP_URL = f"{CISCO_BASE}/api/v1/inspect/mcp"
-
-
-@contextmanager
-def _patch_inspection_post(g: CiscoAIDefenseGuardrail, post_mock: Any):
-    async def _send(request: Request, **kwargs: Any) -> Response:
-        return await post_mock(
-            url=str(request.url),
-            headers=request.headers,
-            json=json.loads(request.content.decode("utf-8")),
-            follow_redirects=kwargs.get("follow_redirects"),
-        )
-
-    with patch.object(g.async_handler.client, "send", new=_send):
-        yield post_mock
-
-
-def _mock_inspect_response(
-    json_body: dict, *, status: int = 200, url: str = CHAT_URL
-) -> Response:
-    return Response(
-        status_code=status,
-        json=json_body,
-        request=Request(method="POST", url=url),
-    )
-
-
-def _safe_response(url: str = CHAT_URL) -> Response:
-    return _mock_inspect_response(
-        {
-            "is_safe": True,
-            "classifications": [],
-            "severity": "NONE_SEVERITY",
-            "rules": [],
-            "action": "allow",
-        },
-        url=url,
-    )
-
-
-def _violation_response(url: str = CHAT_URL) -> Response:
-    return _mock_inspect_response(
-        {
-            "is_safe": False,
-            "classifications": ["SECURITY_VIOLATION", "PRIVACY_VIOLATION"],
-            "severity": "HIGH",
-            "rules": [
-                {"rule_name": "Prompt Injection"},
-                {"rule_name": "PII", "entity_types": ["Email Address"]},
-            ],
-            "explanation": "Detected jailbreak attempt with PII exfiltration",
-            "event_id": "evt_123",
-            "action": "block",
-        },
-        url=url,
-    )
-
-
-def _mcp_request(name="lookup", args=None, jsonrpc=False, **extra):
-    args = args if args is not None else {}
-    if jsonrpc:
-        return {
-            "jsonrpc": "2.0",
-            "id": "1",
-            "method": "tools/call",
-            "params": {"name": name, "arguments": args},
-            **extra,
-        }
-    return {"mcp_tool_name": name, "mcp_arguments": args, **extra}
-
-
-def _mcp_response(content=None, response_cost=0.0):
-    if content is None:
-        content = [{"type": "text", "text": "ok"}]
-    return SimpleNamespace(
-        mcp_tool_call_response=content,
-        hidden_params=SimpleNamespace(response_cost=response_cost),
-    )
-
-
-def _mcp_result_text(content) -> str:
-    if not content:
-        return ""
-    item = content[0] if isinstance(content, list) else content
-    return getattr(item, "text", None) or item.get("text", "")
-
-
-def _chat_request_tool_call_args(arguments: str) -> dict:
-    return {
-        "messages": [
-            {
-                "role": "assistant",
-                "content": None,
-                "tool_calls": [
-                    {
-                        "id": "call_1",
-                        "type": "function",
-                        "function": {
-                            "name": "send_data",
-                            "arguments": arguments,
-                        },
-                    }
-                ],
-            }
-        ]
-    }
-
-
-def _chat_request_function_call_args(arguments: str) -> dict:
-    return {
-        "messages": [
-            {
-                "role": "assistant",
-                "content": None,
-                "function_call": {
-                    "name": "exfil",
-                    "arguments": arguments,
-                },
-            }
-        ]
-    }
-
-
-def _redact_response(
-    *,
-    sanitized_text=None,
-    sanitized_messages=None,
-    sanitized_mcp_arguments=None,
-    sanitized_payload=None,
-    classifications=("PRIVACY_VIOLATION",),
-    rules=({"rule_name": "PII"},),
-    severity="HIGH",
-    url=CHAT_URL,
-):
-    """Build a Cisco verdict that asks for redact with the supplied sanitized payload."""
-    body = {
-        "is_safe": False,
-        "classifications": list(classifications),
-        "severity": severity,
-        "rules": list(rules),
-        "action": "redact",
-    }
-    if sanitized_text is not None:
-        body["sanitized_text"] = sanitized_text
-    if sanitized_messages is not None:
-        body["sanitized_messages"] = sanitized_messages
-    if sanitized_mcp_arguments is not None:
-        body["sanitized_mcp_arguments"] = sanitized_mcp_arguments
-    if sanitized_payload is not None:
-        body["sanitized_payload"] = sanitized_payload
-    return _mock_inspect_response(body, url=url)
-
-
-def _responses_api_response(text, role="assistant"):
-    """Build a ResponsesAPIResponse with a single message output."""
-    from litellm.types.llms.openai import ResponsesAPIResponse
-    from litellm.types.responses.main import GenericResponseOutputItem, OutputText
-
-    return ResponsesAPIResponse(
-        id="resp_1",
-        created_at=0,
-        output=[
-            GenericResponseOutputItem(
-                type="message",
-                id="msg_1",
-                status="completed",
-                role=role,
-                content=[OutputText(type="output_text", text=text, annotations=[])],
-            )
-        ],
-        parallel_tool_calls=False,
-        tool_choice=None,
-        tools=None,
-        top_p=None,
-        usage=None,
-    )
-
-
-def _make_guardrail(
-    inspection_type="chat",
-    event_hook="pre_call",
-    *,
-    name="t",
-    api_key="x",
-    default_on=True,
-    **kwargs,
-):
-    """Construct a Cisco AI Defense guardrail with sensible test defaults."""
-    return CiscoAIDefenseGuardrail(
-        guardrail_name=name,
-        api_key=api_key,
-        inspection_type=inspection_type,
-        event_hook=event_hook,
-        default_on=default_on,
-        **kwargs,
-    )
 
 
 def test_cisco_ai_defense_config_via_init_v2_chat(monkeypatch):
@@ -264,27 +55,6 @@ def test_cisco_ai_defense_config_via_init_v2_chat(monkeypatch):
                     "guardrail": "cisco_ai_defense",
                     "mode": "pre_call",
                     "default_on": True,
-                },
-            }
-        ],
-        config_file_path="",
-    )
-
-
-def test_cisco_ai_defense_config_via_init_v2_mcp(monkeypatch):
-    """init_guardrails_v2 accepts an mcp-mode cisco_ai_defense guardrail."""
-    monkeypatch.setenv("CISCO_AI_DEFENSE_API_KEY", "test-key")
-    litellm.guardrail_name_config_map = {}
-
-    init_guardrails_v2(
-        all_guardrails=[
-            {
-                "guardrail_name": "cisco-mcp",
-                "litellm_params": {
-                    "guardrail": "cisco_ai_defense",
-                    "mode": "pre_mcp_call",
-                    "default_on": True,
-                    "optional_params": {"inspection_type": "mcp"},
                 },
             }
         ],
@@ -345,18 +115,6 @@ def test_init_registers_on_both_callbacks_and_success_callback(monkeypatch):
         "litellm_logging.async_post_mcp_tool_call_hook will skip it, "
         "so MCP responses will never be scanned."
     )
-
-
-def _find_callback(name):
-    """Pull the freshly-registered Cisco callback off litellm.callbacks."""
-    from litellm.proxy.guardrails.guardrail_hooks.cisco_ai_defense import (
-        CiscoAIDefenseGuardrail,
-    )
-
-    for cb in litellm.callbacks:
-        if isinstance(cb, CiscoAIDefenseGuardrail) and cb.guardrail_name == name:
-            return cb
-    raise AssertionError(f"Cisco guardrail {name!r} not in litellm.callbacks")
 
 
 class TestCiscoAIDefenseFlattenedConfig:
@@ -717,540 +475,6 @@ class TestCiscoAIDefenseChatMode:
                     user_api_key_dict=UserAPIKeyAuth(),
                     response=response,
                 )
-
-
-class TestCiscoAIDefenseMCPMode:
-    @pytest.mark.asyncio
-    async def test_mcp_mode_inspects_mcp_request(self):
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        data = _mcp_request(
-            name="send_email", args={"to": "x@y.com"}, litellm_call_id="call-1"
-        )
-        post_mock = AsyncMock(return_value=_safe_response(url=MCP_URL))
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_pre_call_hook(
-                user_api_key_dict=UserAPIKeyAuth(),
-                cache=DualCache(),
-                data=data,
-                call_type="mcp_call",
-            )
-        assert result == data
-        assert post_mock.call_args.kwargs["url"] == MCP_URL
-        assert post_mock.call_args.kwargs["follow_redirects"] is False
-        sent_payload = post_mock.call_args.kwargs["json"]
-        assert sent_payload["jsonrpc"] == "2.0"
-        assert sent_payload["method"] == "tools/call"
-        assert sent_payload["params"]["name"] == "send_email"
-        assert sent_payload["params"]["arguments"] == {"to": "x@y.com"}
-        assert "request" not in sent_payload
-        assert "metadata" not in sent_payload
-        assert "config" not in sent_payload
-
-    @pytest.mark.asyncio
-    async def test_mcp_mode_blocks_violation(self):
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        data = _mcp_request(name="leak_secrets", args={"target": "evil"})
-        with _patch_inspection_post(
-            g, AsyncMock(return_value=_violation_response(url=MCP_URL))
-        ):
-            with pytest.raises(HTTPException) as exc:
-                await g.async_pre_call_hook(
-                    user_api_key_dict=UserAPIKeyAuth(),
-                    cache=DualCache(),
-                    data=data,
-                    call_type="mcp_call",
-                )
-        assert exc.value.detail["surface"] == "mcp"
-
-    @pytest.mark.asyncio
-    async def test_mcp_mode_skips_chat_traffic(self):
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        data = {"messages": [{"role": "user", "content": "hello"}]}
-        post_mock = AsyncMock()
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_pre_call_hook(
-                user_api_key_dict=UserAPIKeyAuth(),
-                cache=DualCache(),
-                data=data,
-                call_type="completion",
-            )
-        assert result == data
-        post_mock.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_mcp_mode_inspects_jsonrpc_envelope(self):
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        data = _mcp_request(name="do_thing", args={"x": 1}, jsonrpc=True, id="abc")
-        post_mock = AsyncMock(return_value=_safe_response(url=MCP_URL))
-        with _patch_inspection_post(g, post_mock):
-            await g.async_pre_call_hook(
-                user_api_key_dict=UserAPIKeyAuth(),
-                cache=DualCache(),
-                data=data,
-                call_type="mcp_call",
-            )
-        sent_payload = post_mock.call_args.kwargs["json"]
-        assert sent_payload["jsonrpc"] == "2.0"
-        assert sent_payload["id"] == "abc"
-        assert sent_payload["params"]["name"] == "do_thing"
-        assert sent_payload["params"]["arguments"] == {"x": 1}
-
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_inspects_tool_output(self):
-        """The post-MCP hook is what the proxy actually calls for MCP responses.
-
-        post_call is not registered for mcp-mode guardrails, so MCP tool
-        output must go through async_post_mcp_tool_call_hook.
-
-        """
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook=["pre_mcp_call", "during_mcp_call"],
-            default_on=True,
-        )
-
-        response_obj = _mcp_response(
-            SimpleNamespace(
-                content=[{"type": "text", "text": "Here is the secret API key abc123"}]
-            )
-        )
-
-        post_mock = AsyncMock(return_value=_safe_response(url=MCP_URL))
-        kwargs = {
-            "name": "lookup_secret",
-            "arguments": {"key": "production"},
-            "mcp_server_name": "vault",
-            "litellm_call_id": "call-42",
-        }
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs=kwargs,
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert result is None
-        assert post_mock.called
-        assert post_mock.call_args.kwargs["url"] == MCP_URL
-        sent_payload = post_mock.call_args.kwargs["json"]
-        assert sent_payload["jsonrpc"] == "2.0"
-        assert sent_payload["id"] == "call-42"
-        assert sent_payload["method"] == "tools/call"
-        assert sent_payload["params"] == {
-            "name": "lookup_secret",
-            "arguments": {"key": "production"},
-        }
-        assert sent_payload["result"]["content"][0]["text"] == (
-            "Here is the secret API key abc123"
-        )
-        assert "request" not in sent_payload
-        assert "metadata" not in sent_payload
-
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_blocks_violation(self):
-        """A violation verdict from the MCP response scan must enforce a block.
-
-        The litellm post-MCP dispatcher wraps every callback in
-        ``try: ... except Exception`` and logs as non-blocking, so raising
-        an ``HTTPException`` here would be silently swallowed and the
-        unsafe tool output would still reach the caller. The hook must
-        therefore return a non-None ``MCPPostCallResponseObject`` whose
-        ``mcp_tool_call_response`` replaces the original tool output with
-        a synthetic violation message — the only blocking lever the
-        dispatcher exposes (litellm_logging.py
-        ``_parse_post_mcp_call_hook_response``).
-        """
-        from litellm.types.mcp import MCPPostCallResponseObject
-
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook=["pre_mcp_call", "during_mcp_call"],
-            default_on=True,
-        )
-        response_obj = _mcp_response(
-            SimpleNamespace(content=[{"type": "text", "text": "leaked"}])
-        )
-
-        post_mock = AsyncMock(return_value=_violation_response(url=MCP_URL))
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs={"name": "leak", "arguments": {}},
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert result is not None, (
-            "MCP response block was silently dropped — the litellm "
-            "dispatcher swallows raised exceptions, so the hook must "
-            "return a non-None MCPPostCallResponseObject to enforce a block."
-        )
-        assert isinstance(result, MCPPostCallResponseObject)
-        replacement = result.mcp_tool_call_response
-        assert len(replacement) == 1
-        text = _mcp_result_text(replacement)
-        assert "Blocked by Cisco AI Defense" in text
-        assert "evt_123" in text
-        assert "SECURITY_VIOLATION" in text
-
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_skipped_in_chat_mode(self):
-        """Chat-mode guardrails must NOT scan MCP tool output, even via this hook."""
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="chat",
-            event_hook="pre_call",
-            default_on=True,
-        )
-        response_obj = _mcp_response(
-            SimpleNamespace(content=[{"type": "text", "text": "hi"}])
-        )
-
-        post_mock = AsyncMock()
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs={"name": "tool", "arguments": {}},
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-        assert result is None
-        post_mock.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_post_call_skipped_for_mcp_mode_guardrail(self):
-        """Even if dispatched, the chat post_call hook must no-op in mcp mode."""
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        data = {"messages": [{"role": "user", "content": "hi"}]}
-        response = _make_model_response_with_content("fine")
-
-        post_mock = AsyncMock()
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_call_success_hook(
-                data=data,
-                user_api_key_dict=UserAPIKeyAuth(),
-                response=response,
-            )
-        assert result is response
-        post_mock.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_runs_with_pre_mcp_call_only(self):
-        """Per product decision: ``pre_mcp_call`` means "guard the MCP
-        call" — request AND response. Response scanning must fire even
-        when only ``pre_mcp_call`` is configured (no explicit
-        ``during_mcp_call``).
-        """
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        response_obj = _mcp_response(
-            SimpleNamespace(
-                content=[{"type": "text", "text": "would have been scanned"}]
-            )
-        )
-
-        post_mock = AsyncMock(return_value=_safe_response(url=MCP_URL))
-        with _patch_inspection_post(g, post_mock):
-            await g.async_post_mcp_tool_call_hook(
-                kwargs={"name": "lookup", "arguments": {}},
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert post_mock.called, (
-            "MCP response scan was skipped when only ``pre_mcp_call`` "
-            "was configured. Per product decision, pre_mcp_call means "
-            "'guard the MCP call' — request AND response."
-        )
-
-    @pytest.mark.parametrize(
-        "cisco_response_kind,expected_block",
-        [("safe", False), ("violation", True)],
-    )
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_handles_raw_list_content(
-        self, cisco_response_kind, expected_block
-    ):
-        """Regression: the production post-MCP dispatcher hands us a raw
-        list (see ``MCPPostCallResponseObject.mcp_tool_call_response``
-        typing). Inspect must fire for the raw-list shape, and a
-        violation verdict must produce a synthetic ``MCPPostCallResponseObject``.
-        """
-        from litellm.types.mcp import MCPPostCallResponseObject
-
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook=["pre_mcp_call", "during_mcp_call"],
-            default_on=True,
-        )
-
-        text_content = (
-            "exfiltrated data: ..."
-            if cisco_response_kind == "violation"
-            else "Here is the secret API key abc123"
-        )
-        response_obj = _mcp_response([{"type": "text", "text": text_content}])
-
-        cisco_resp = (
-            _violation_response(url=MCP_URL)
-            if cisco_response_kind == "violation"
-            else _safe_response(url=MCP_URL)
-        )
-        post_mock = AsyncMock(return_value=cisco_resp)
-        kwargs = {
-            "name": "leak" if expected_block else "lookup_secret",
-            "arguments": {"key": "production"} if not expected_block else {},
-            "mcp_server_name": "vault",
-            "litellm_call_id": "call-raw-list",
-        }
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs=kwargs,
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert post_mock.called, (
-            "MCP response inspect was silently skipped for raw-list "
-            "shape — _normalize_mcp_response failed."
-        )
-        assert post_mock.call_args.kwargs["url"] == MCP_URL
-
-        if expected_block:
-            assert isinstance(result, MCPPostCallResponseObject)
-            replacement = result.mcp_tool_call_response
-            assert len(replacement) == 1
-            assert "Blocked by Cisco AI Defense" in _mcp_result_text(replacement)
-        else:
-            sent_payload = post_mock.call_args.kwargs["json"]
-            assert sent_payload["jsonrpc"] == "2.0"
-            assert sent_payload["id"] == "call-raw-list"
-            assert sent_payload["method"] == "tools/call"
-            assert sent_payload["params"] == {
-                "name": "lookup_secret",
-                "arguments": {"key": "production"},
-            }
-            assert sent_payload["result"]["content"][0]["text"] == text_content
-            assert result is None
-
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_through_real_logging_wrapper(self):
-        """Real dispatcher shape must send content and structuredContent."""
-        from mcp.types import CallToolResult, TextContent
-
-        from litellm.types.mcp import MCPPostCallResponseObject
-
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook=["pre_mcp_call", "during_mcp_call"],
-            default_on=True,
-        )
-
-        real_result = CallToolResult(
-            content=[TextContent(type="text", text="leak 9045629876")],
-            structuredContent={"patient": {"ssn": "123-45-6789"}},
-            isError=False,
-        )
-        wrapped = MCPPostCallResponseObject(
-            mcp_tool_call_response=real_result,
-            hidden_params={},
-        )
-
-        assert isinstance(wrapped.mcp_tool_call_response, list)
-        assert all(
-            isinstance(item, tuple) and len(item) == 2
-            for item in wrapped.mcp_tool_call_response
-        ), (
-            "Pydantic coercion shape changed — update the normalizer to "
-            "match the new wire format."
-        )
-
-        post_mock = AsyncMock(return_value=_safe_response(url=MCP_URL))
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs={
-                    "name": "leak_tool",
-                    "arguments": {},
-                    "mcp_server_name": "vault",
-                    "litellm_call_id": "real-wire-call",
-                },
-                response_obj=wrapped,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert post_mock.called, (
-            "Inspect API not called for real CallToolResult shape — "
-            "_normalize_mcp_response failed to handle Pydantic's "
-            "iterated-BaseModel coercion."
-        )
-        assert post_mock.call_args.kwargs["url"] == MCP_URL
-        sent_payload = post_mock.call_args.kwargs["json"]
-        content_items = sent_payload["result"]["content"]
-
-        assert len(content_items) == 1, (
-            f"expected exactly 1 content item from the real "
-            f"CallToolResult.content list, got {len(content_items)}: "
-            f"{content_items!r}"
-        )
-        assert content_items[0].get("text") == "leak 9045629876", (
-            f"Cisco wire payload missed the real tool text; got "
-            f"{content_items[0]!r}. This means the Pydantic-coerced "
-            f"(field_name, value) tuple shape was serialized as text "
-            f"content instead of being unwrapped to find the inner "
-            f"``content`` field."
-        )
-        assert content_items[0].get("type") == "text"
-        assert sent_payload["result"]["structuredContent"] == {
-            "patient": {"ssn": "123-45-6789"}
-        }
-        assert sent_payload["result"]["isError"] is False
-        assert sent_payload["id"] == "real-wire-call"
-        assert sent_payload["method"] == "tools/call"
-        assert sent_payload["params"] == {"name": "leak_tool", "arguments": {}}
-        assert result is None
-
-    @pytest.mark.asyncio
-    async def test_mcp_response_hook_uses_standard_logging_tool_metadata(self):
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        response_obj = _mcp_response([{"type": "text", "text": "tool output"}])
-
-        post_mock = AsyncMock(return_value=_safe_response(url=MCP_URL))
-        with _patch_inspection_post(g, post_mock):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs={
-                    "litellm_call_id": "metadata-call",
-                    "mcp_tool_call_metadata": {
-                        "name": "lookup_secret",
-                        "arguments": {"key": "production"},
-                        "mcp_server_name": "vault",
-                    },
-                },
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert result is None
-        sent_payload = post_mock.call_args.kwargs["json"]
-        assert sent_payload["method"] == "tools/call"
-        assert sent_payload["params"] == {
-            "name": "lookup_secret",
-            "arguments": {"key": "production"},
-        }
-        assert sent_payload["result"]["content"][0]["text"] == "tool output"
-
-
-def _make_streaming_chunks(parts):
-    """Build a list of OpenAI-shape streaming chunks from text segments."""
-    chunks = []
-    for i, part in enumerate(parts):
-        chunks.append(
-            ModelResponseStream(
-                id="resp_1",
-                choices=[
-                    StreamingChoices(
-                        delta=Delta(content=part, role="assistant" if i == 0 else None),
-                        finish_reason="stop" if i == len(parts) - 1 else None,
-                        index=0,
-                    )
-                ],
-                created=1234567890,
-                model="gpt-4",
-                object="chat.completion.chunk",
-            )
-        )
-    return chunks
-
-
-async def _aiter(items):
-    for item in items:
-        yield item
-
-
-async def _streaming_setup(
-    g,
-    chunks,
-    cisco_response=None,
-    upstream=None,
-    request_data=None,
-    post_mock=None,
-):
-    """Run the streaming hook with the standard patch/iterate boilerplate.
-
-    Returns ``(received_chunks, post_mock)``. Caller picks the post mock
-    flavor (``AsyncMock(return_value=cisco_response)`` by default, or any
-    custom mock supplied via ``post_mock``).
-    """
-    if post_mock is None:
-        post_mock = (
-            AsyncMock(return_value=cisco_response) if cisco_response else AsyncMock()
-        )
-    stream_source = upstream if upstream is not None else _aiter(chunks)
-    if request_data is None:
-        request_data = {"messages": [{"role": "user", "content": "hi"}]}
-    received: list = []
-    with _patch_inspection_post(g, post_mock):
-        async for chunk in g.async_post_call_streaming_iterator_hook(
-            user_api_key_dict=UserAPIKeyAuth(),
-            response=stream_source,
-            request_data=request_data,
-        ):
-            received.append(chunk)
-    return received, post_mock
 
 
 class TestCiscoAIDefenseResponsesAPIOutput:
@@ -2445,89 +1669,6 @@ class TestCiscoAIDefenseEnabledRulesPydanticShape:
         assert result["entity_types"] == ["SSN"]
 
 
-class TestCiscoAIDefenseRedactListShape:
-    """MCP response redaction handles list-shaped tool output."""
-
-    @staticmethod
-    def _violation_with_redact_response(text: str = "[REDACTED tool output]"):
-        """Cisco verdict requesting redact with a sanitized_text rewrite."""
-        return _mock_inspect_response(
-            {
-                "is_safe": False,
-                "classifications": ["PRIVACY_VIOLATION"],
-                "severity": "HIGH",
-                "rules": [{"rule_name": "PII", "entity_types": ["SSN"]}],
-                "explanation": "PII detected, redaction available",
-                "event_id": "evt_redact_1",
-                "action": "redact",
-                "sanitized_text": text,
-            },
-            url=MCP_URL,
-        )
-
-    @staticmethod
-    def _raw_list_factory():
-        original_content = [{"type": "text", "text": "Your SSN is 123-45-6789."}]
-        return original_content, lambda: original_content[0]["text"]
-
-    @staticmethod
-    def _pydantic_tuple_list_factory():
-        from mcp.types import TextContent
-
-        inner_content = [TextContent(type="text", text="SSN: 123-45-6789")]
-        tuples_list = [
-            ("meta", None),
-            ("content", inner_content),
-            ("structuredContent", {"patient": {"ssn": "123-45-6789"}}),
-            ("isError", False),
-        ]
-        return tuples_list, lambda: inner_content[0].text
-
-    @pytest.mark.parametrize(
-        "factory_name",
-        ["_raw_list_factory", "_pydantic_tuple_list_factory"],
-    )
-    @pytest.mark.asyncio
-    async def test_redact_rewrites_mcp_response_list_shape(self, factory_name):
-        """Raw content lists and Pydantic tuple lists are both rewritten."""
-
-        from litellm.types.mcp import MCPPostCallResponseObject
-
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="t",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook=["pre_mcp_call", "during_mcp_call"],
-            default_on=True,
-        )
-
-        content, get_text = getattr(self, factory_name)()
-        response_obj = _mcp_response(content)
-
-        with _patch_inspection_post(
-            g, AsyncMock(return_value=self._violation_with_redact_response())
-        ):
-            result = await g.async_post_mcp_tool_call_hook(
-                kwargs={"name": "leak", "arguments": {}},
-                response_obj=response_obj,
-                start_time=datetime.now(),
-                end_time=datetime.now(),
-            )
-
-        assert result is None or not isinstance(result, MCPPostCallResponseObject), (
-            f"Redact silently fell through to block for {factory_name}. "
-            f"result={result!r}"
-        )
-        assert get_text() == "[REDACTED tool output]", (
-            f"Redact silently failed for {factory_name}; original text "
-            f"not rewritten."
-        )
-        if factory_name == "_pydantic_tuple_list_factory":
-            structured_content = dict(content)["structuredContent"]
-            assert structured_content == {"result": "[REDACTED tool output]"}
-            assert "123-45-6789" not in json.dumps(structured_content)
-
-
 class TestCiscoAIDefenseResponsesAPIBypass:
     """Responses API ``input`` shapes must be flattened for scanning."""
 
@@ -3260,72 +2401,6 @@ class TestCiscoAIDefenseSurfaceBypass:
             post_mock.assert_not_called()
 
 
-class TestCiscoAIDefenseMCPBlockingContract:
-    """MCP response blocks survive non-enforcing dispatcher paths."""
-
-    @pytest.mark.asyncio
-    async def test_block_response_survives_dispatcher_contract(self):
-        """Block must mutate the original tool result and return a synthetic response."""
-        from litellm.litellm_core_utils.litellm_logging import Logging
-        from litellm.types.mcp import MCPPostCallResponseObject
-        from mcp.types import CallToolResult, TextContent
-
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="cisco-mcp",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook=["pre_mcp_call", "during_mcp_call"],
-            default_on=True,
-        )
-        raw_response = CallToolResult(
-            content=[TextContent(type="text", text="exfiltrated")],
-            structuredContent={"result": "exfiltrated"},
-            isError=False,
-        )
-        response_obj = MCPPostCallResponseObject(
-            mcp_tool_call_response=raw_response,
-            hidden_params={},
-        )
-
-        post_mock = AsyncMock(return_value=_violation_response(url=MCP_URL))
-        captured: Dict[str, Any] = {}
-        with _patch_inspection_post(g, post_mock):
-            try:
-                captured["result"] = await g.async_post_mcp_tool_call_hook(
-                    kwargs={
-                        "name": "leak",
-                        "arguments": {},
-                        "original_response": raw_response,
-                    },
-                    response_obj=response_obj,
-                    start_time=datetime.now(),
-                    end_time=datetime.now(),
-                )
-            except Exception as e:
-                captured["swallowed"] = repr(e)
-
-        assert "swallowed" not in captured, (
-            f"async_post_mcp_tool_call_hook raised — the litellm "
-            f"dispatcher would swallow this and the block would be lost. "
-            f"Got: {captured.get('swallowed')}"
-        )
-        result = captured["result"]
-        assert isinstance(result, MCPPostCallResponseObject), (
-            "Hook must keep returning a MCPPostCallResponseObject for "
-            "dispatcher paths that do honor returned replacements."
-        )
-        assert raw_response.isError is True
-        assert "Blocked by Cisco AI Defense" in raw_response.content[0].text
-        assert raw_response.structuredContent is not None
-        assert "Blocked by Cisco AI Defense" in raw_response.structuredContent["result"]
-        assert "exfiltrated" not in raw_response.structuredContent["result"]
-        logging_stub = Logging.__new__(Logging)
-        logging_stub.model_call_details = {}
-        parsed = logging_stub._parse_post_mcp_call_hook_response(response=result)
-        assert parsed is not None
-        assert "Blocked by Cisco AI Defense" in _mcp_result_text(parsed)
-
-
 class TestCiscoAIDefenseEventTypeDirection:
     """Input and output scans log under distinct event types."""
 
@@ -3650,138 +2725,6 @@ class TestCiscoAIDefenseRedactAction:
                     call_type="completion",
                 )
         assert exc.value.status_code == 400
-
-
-class TestCiscoAIDefenseJsonRpcSuccessEnvelope:
-    """Regression: Cisco's /inspect/mcp endpoint wraps the verdict under
-    ``result`` (JSON-RPC). The handler must unwrap it; otherwise unsafe
-    MCP traffic was silently passing through.
-
-    Captured shape from a real Cisco AI Defense response:
-
-        {
-          "jsonrpc": "2.0",
-          "id": 3,
-          "result": {
-            "is_safe": false,
-            "action": "Block",
-            "classifications": [],
-            "rules": [
-              {"rule_name": "PII", "classification": "NONE_VIOLATION"}
-            ],
-            "event_id": "..."
-          }
-        }
-    """
-
-    @staticmethod
-    def _cisco_mcp_envelope(*, is_safe: bool, action: str = "Block") -> Response:
-        return _mock_inspect_response(
-            {
-                "jsonrpc": "2.0",
-                "id": 3,
-                "result": {
-                    "is_safe": is_safe,
-                    "action": action,
-                    "classifications": [],
-                    "rules": [
-                        {
-                            "rule_name": "PII",
-                            "rule_id": 0,
-                            "entity_types": [],
-                            "classification": "NONE_VIOLATION",
-                        }
-                    ],
-                    "event_id": "645d9d22-b016-47e0-a12c-9d587fb11c57",
-                    "detected_pii": [],
-                },
-            },
-            url=MCP_URL,
-        )
-
-    @pytest.mark.parametrize(
-        "is_safe,action,should_block",
-        [
-            (False, "Block", True),
-            (True, "Allow", False),
-            (False, "Allow", False),
-            (True, "Block", True),
-        ],
-    )
-    @pytest.mark.asyncio
-    async def test_mcp_jsonrpc_envelope_respects_verdict(
-        self, is_safe, action, should_block
-    ):
-        g = CiscoAIDefenseGuardrail(
-            guardrail_name="cisco-mcp",
-            api_key="x",
-            inspection_type="mcp",
-            event_hook="pre_mcp_call",
-            default_on=True,
-        )
-        data = _mcp_request(
-            name="ask_question",
-            args={
-                "repoName": "facebook/react",
-                "question": "What is React Fiber 9045629876?",
-            },
-        )
-        with _patch_inspection_post(
-            g,
-            AsyncMock(
-                return_value=self._cisco_mcp_envelope(is_safe=is_safe, action=action)
-            ),
-        ):
-            if should_block:
-                with pytest.raises(HTTPException) as exc:
-                    await g.async_pre_call_hook(
-                        user_api_key_dict=UserAPIKeyAuth(),
-                        cache=DualCache(),
-                        data=data,
-                        call_type="mcp_call",
-                    )
-                assert exc.value.status_code == 400
-                assert exc.value.detail["surface"] == "mcp"
-                assert (
-                    exc.value.detail["event_id"]
-                    == "645d9d22-b016-47e0-a12c-9d587fb11c57"
-                )
-            else:
-                result = await g.async_pre_call_hook(
-                    user_api_key_dict=UserAPIKeyAuth(),
-                    cache=DualCache(),
-                    data=data,
-                    call_type="mcp_call",
-                )
-                assert result == data
-
-    @pytest.mark.parametrize(
-        "verdict,expected",
-        [
-            (
-                {
-                    "is_safe": False,
-                    "classifications": ["SECURITY_VIOLATION"],
-                    "action": "block",
-                },
-                "passthrough",
-            ),
-            (
-                {
-                    "jsonrpc": "2.0",
-                    "id": 1,
-                    "result": {"is_safe": False, "action": "Block"},
-                },
-                {"is_safe": False, "action": "Block"},
-            ),
-        ],
-    )
-    def test_unwrap_verdict_envelope(self, verdict, expected):
-        unwrapped = CiscoAIDefenseGuardrail._unwrap_verdict_envelope(verdict)
-        if expected == "passthrough":
-            assert unwrapped is verdict
-        else:
-            assert unwrapped == expected
 
 
 class TestCiscoAIDefenseJsonRpcError:
